requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
attrs==25.4.0
click==8.3.1
discord==2.3.2
discord.py==2.7.1
frozenlist==1.8.0
idna==3.11
irc==20.5.0
jaraco.collections==5.2.1
jaraco.context==6.1.1
jaraco.functools==4.4.0
jaraco.logging==3.4.0
jaraco.stream==3.0.4
jaraco.text==4.2.0
markdown-it-py==4.0.0
mdurl==0.1.2
more-itertools==10.8.0
multidict==6.7.1
propcache==0.4.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
rich==14.3.3
shellingham==1.5.4
six==1.17.0
tempora==5.8.1
typer==0.24.1
typer-slim==0.24.0
yarl==1.23.0
joblib==1.5.3
numpy==2.4.3
scikit-learn==1.8.0
scipy==1.17.1
threadpoolctl==3.6.0
huggingface_hub==1.7.2
sentence-transformers==5.3.0
